Delete unused class.  Do not run unneeded diff when comparing to base revision. Issue #: 1076

+++ org.tigris.subversion.subclipse.ui/src/org/tigris/subversion/subclipse/ui/actions/CompareWithRemoteAction.java
@@ -58,11 +58,15 @@ public abstract class CompareWithRemoteAction extends WorkbenchWindowAction {
 				public void run(IProgressMonitor monitor) {
 					try {
 						if (refresh) localResource.getResource().refreshLocal(Depth.immediates, monitor);				
-						File file = File.createTempFile("revision", ".diff");
-						file.deleteOnExit();
-						ShowDifferencesAsUnifiedDiffOperationWC operation = new ShowDifferencesAsUnifiedDiffOperationWC(getTargetPart(), localResource.getFile(), localResource.getUrl(), SVNRevision.HEAD, file);						
-						operation.setGraphicalCompare(true);
-						operation.run();
+						File file = null;
+						ShowDifferencesAsUnifiedDiffOperationWC operation = null;
+						if (SVNRevision.HEAD.equals(revision)) {
+							file = File.createTempFile("revision", ".diff");
+							file.deleteOnExit();
+							operation = new ShowDifferencesAsUnifiedDiffOperationWC(getTargetPart(), localResource.getFile(), localResource.getUrl(), SVNRevision.HEAD, file);						
+							operation.setGraphicalCompare(true);
+							operation.run();
+						}
 						SVNLocalCompareInput compareInput = new SVNLocalCompareInput(localResource, revision);
 						compareInput.setDiffFile(file);
 						compareInput.setDiffOperation(operation);
